% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDBB.R
\name{multiple_GSB}
\alias{multiple_GSB}
\title{Title}
\usage{
multiple_GSB(a, iteration, burn_in, bic_row_labels, number_bicluster)
}
\arguments{
\item{a}{input data matrix}

\item{iteration}{iteration Number of initial iterations}

\item{burn_in}{Number of burnings}

\item{bic_row_labels}{Labels for rows in the patterns found, all with initial values of 0}

\item{number_bicluster}{Number of preset patterns}
}
\description{
Title
}
\examples{
library(BDBB)
number_bicluster <- 6
bic_row_labels <- c(rep(0,dim(mixture)[1]))
simulate_multiple_GSB_object <- multiple_GSB(a = mixture,iteration = 2000,burn_in = 500,bic_row_labels = bic_row_labels,number_bicluster = number_bicluster)

}
